Since the owner of these two files already left the chair I fixed the compiler warnings caused in these files.
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@13044 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_mat/charmm.cpp
@@ -127,8 +127,8 @@ void MAT::CHARMM::Unpack(const vector<char>& data) {
 	params_ = NULL;
     }
 
-    if (position != (int) data.size())
-	dserror("Mismatch in size of data %d <-> %d", (int) data.size(), position);
+    if (position != data.size())
+      dserror("Mismatch in size of data %d <-> %d", data.size(), position);
 }
